// product-wizard-spec/canvas wizard spec reference/coverageVariant/coverage-factors/model/coverage-factors.model.ts
export interface CoverageFactorsLabels {
  back: string;
  next: string;
  saveAndExit: string;
  title: string;
  active: string;
  inActive: string;
  ageBand: string;
  gender: string;
  createSucessMessage: string;
  createErrorMessage: string;
  updateSucessMessage: string;
  updateErrorMessage: string;
  errorMessage: string;
  ageError: string;
  formError: string;
  cancelButton: string;
  okButton: string;
  modalTitle: string;
  modalText: string;
}

export enum CoverageFactorsEnum {
  AGE = 'AGE',
  GENDER = 'GENDER',
}